First release of li3_jade everything works but here are some notes:
The basic set up was to register a new media type with li3 which the
bootstrap file does.  Then create the compiler and renderer classes.
I used a two step process to get the php to be evaled: the template is
cached and passed to jade parser, the resulting php is written to a
cached file as well and that is what is added to the output buffer.